requirements.txt
# FastAPI and ASGI server
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.1
pydantic-settings>=2.1.0

# Data processing
pandas>=2.1.3
numpy>=1.26.2
swifter>=1.4.0
openpyxl>=3.1.2  # for Excel file support

# Google Cloud
google-cloud-storage>=2.13.0

# Async support
aiofiles>=23.2.1
asyncio>=3.4.3

# Utilities
python-dotenv>=1.0.0
filelock>=3.13.1

# Logging
logging>=0.4.9.6

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2  # for async HTTP client

# Development
black>=23.11.0  # code formatting
isort>=5.12.0  # import sorting
flake8>=6.1.0  # code linting
mypy>=1.7.1  # type checking

# CORS
fastapi-cors>=0.0.6

# Compression
python-snappy>=0.6.1